moved plate location getter and setter to PlateJoint from PlateJointCandidate
